[lib] turn static class "enums" into integer constants
The static class "enum" pattern need to allocate
one chunk of heap for each enum element,
thus using a lot of runtime memory.
Integers on the other hand can be allocated on the stack
and only need RAM when needed. Thus this should
be much better resource wise.
On the down side, this gives up type safety.